+++ src/main/java/io/gravitee/repository/api/management/ApiKeyRepository.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.repository.api;
+package io.gravitee.repository.api.management;
 
 import io.gravitee.repository.exceptions.TechnicalException;
-import io.gravitee.repository.model.Api;
-import io.gravitee.repository.model.ApiKey;
-import io.gravitee.repository.model.Application;
+import io.gravitee.repository.model.management.Api;
+import io.gravitee.repository.model.management.ApiKey;
+import io.gravitee.repository.model.management.Application;
 
 import java.util.Optional;
 import java.util.Set;
