+++ src/main/java/io/gravitee/repository/api/management/TeamMembershipRepository.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.repository.api;
+package io.gravitee.repository.api.management;
 
 import io.gravitee.repository.exceptions.TechnicalException;
-import io.gravitee.repository.model.Member;
-import io.gravitee.repository.model.Team;
+import io.gravitee.repository.model.management.Member;
+import io.gravitee.repository.model.management.Team;
 
 import java.util.Set;
