+++ src/main/java/io/gravitee/repository/api/management/UserRepository.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.repository.api;
+package io.gravitee.repository.api.management;
 
 import java.util.Optional;
 import java.util.Set;
 
 import io.gravitee.repository.exceptions.TechnicalException;
-import io.gravitee.repository.model.User;
+import io.gravitee.repository.model.management.User;
 
 /**
  * @author David BRASSELY (brasseld at gmail.com)
