+++ src/main/java/io/gravitee/repository/model/management/LifecycleState.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.repository.model;
+package io.gravitee.repository.model.management;
 
 /**
  * Api lifecycle states
